pyproject.toml
[project]
name = "latent-ad-qml"
version = "1.0.4"
description = "A quantum machine learning library for latent adaptive models. Fork with enhancements."
authors = [
    { name = "Vasilis Belis (original author)" },
    { name = "Kinga Anna Wozniak (original author)" },
    { name = "Ema Puljak (original author)" }
]
maintainers = [
    { name = "rotiyan (fork maintainer)" }
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
dependencies = [
    "h5py>=3.6.0",
    "joblib>=1.2.0",
    "matplotlib>=3.5",
    "mplhep>=0.3.26",
    "mplhep-data==0.0.3",
    "numpy>=1.21",
    "pandas==1.4.0",
    "scikit-learn==1.1.1",
    "scipy>=1.9",
    "qibo==0.1.10",
    "qiskit==0.36.2",
    "qiskit-aer>=0.10.4",
    "qiskit-ibmq-provider==0.19.1",
    "qiskit-ignis==0.7.1",
    "qiskit-machine-learning==0.4.0",
    "qiskit-terra==0.20.2",
    "tensorflow>=2.6",
    "torch>=1.9.0",
    "pylatexenc==2.10",
    "pyyaml>=6.0",
    "triple-e @ https://github.com/vbelis/triple_e/archive/master.zip"
]
optional-dependencies.docs = [
    "sphinx>=3.0",
    "sphinx-autoapi",
    "numpy>=1.21",
    "pandas>=1.4.0"
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"